requirements.txt
# Core LangChain dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Existing Yona dependencies
openai==1.6.1
httpx==0.25.2
python-dotenv==1.0.0
supabase==2.3.0
flask==3.0.0

# Cryptography for DID management
cryptography==41.0.8

# Additional utilities
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0

# Coral Protocol Official Integration
langchain-mcp-adapters==0.0.10
mcp==1.7.1
httpx-sse==0.4.0
anyio==4.9.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
